requirements.txt
chromadb
google-generativeai
langchain
langchain-text-splitters
pymupdf
python-dotenv
streamlit
pytest
